refactor(core): single-source the wire shapes ui was hand-copying from apps
`ui` may not import `apps` (dependency-guard), so it re-declared the apps
shapes it reads off the wire "verbatim from the frozen contract text". That is
a promise, not a mechanism, and three of the copies had already drifted apart
from their originals in comment and one — `EditResult` — in FIELDS.

Moves the shapes both sides genuinely speak into core, where each has exactly
one definition:

  - `pinComponentName` -> core/app-document.ts, beside `Pin` (it is a pure
    function of `Pin.slot`). It had THREE copies: apps/pins.ts, ui's
    `<Remixable>` wrapper, and ui's wire fixture.
  - `PlacementEntry`, `ReviewStanding` -> core/app-surfaces.ts, a new module
    whose membership rule is stated in its header: apps produces it, ui
    consumes it off the wire.

apps and ui both re-export the moved names from the modules that exported them
before, so no package's public surface changes.

NOT lifted, deliberately: `PinForkResult`. Its own fields match, but its
`edit?: EditResult` does not — apps' `EditResult` carries four fields ui's copy
never grew (`failure`, `graduated`, `box`, `pendingEgress`), and the wire
returns the runtime's result untrimmed (`json(await deps.apps.pins.fork(...))`,
wire/apps.ts). Unifying it would widen `@vendoai/ui`'s published `EditResult`,
which is a contract change, not a refactor. Reconciling that pair is its own
change.

`PlacementRow` stayed in apps: it is the placement STORAGE row, and ui has no
copy of it.
