+++ packages/core/src/app-document.ts
@@ -4,6 +4,7 @@ import { safeErrorMessage } from "./errors.js";
 import { FN_REFERENCE_PATTERN, collectActionReferences } from "./fn-references.js";
 import { VENDO_APP_FORMAT, VENDO_TREE_FORMAT } from "./formats.js";
 import { appIdSchema, isoDateTimeSchema, type AppId, type IsoDateTime } from "./ids.js";
+import { sha256Hex } from "./sha256.js";
 import { TOOL_NAME_PATTERN } from "./tools.js";
 import { validateTree } from "./genui/tree.js";
 import { DEFAULT_TRIGGER_ID, triggerSchema, type Trigger } from "./triggers.js";
@@ -125,6 +126,25 @@ export const pinSchema = z.object({
   slot: z.string(),
   base: z.string(),
 }).passthrough() satisfies z.ZodType<Pin>;
+
+/**
+ * The stable generated-component name a fork of one captured host slot ships
+ * under — a pure function of {@link Pin}'s slot, so it belongs beside the pin.
+ *
+ * It lives in core because BOTH sides of the fork seam need the same answer and
+ * `ui → apps` is not an edge layering allows (dependency-guard): the runtime
+ * writes the node under this name, and the client's in-place mount finds it by
+ * this name. It was hand-copied three times before this (apps' pins.ts, ui's
+ * remixable wrapper, ui's wire fixture).
+ */
+export const pinComponentName = (slot: string): string => {
+  const stem = (slot.match(/[A-Za-z0-9]+/g) ?? [])
+    .map((part) => `${part.slice(0, 1).toUpperCase()}${part.slice(1)}`)
+    .join("") || "Slot";
+  // The hash suffix prevents punctuation-only normalization collisions while
+  // keeping the name a valid generated-component PascalCase identifier.
+  return `Pinned${stem}${sha256Hex(slot).slice(0, 8)}`;
+};
 
 /**
  * execution-v2 — the app's persistent machine. Presence means layer 2+; an app
